third_party/blink/web_tests/platform/generic/fast/dynamic/hover-sibling-reattach-expected.txt

Check that childrenAffectedByHover is intact after a re-attach.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(test, '').backgroundColor is transparent
PASS getComputedStyle(test, '').backgroundColor is black
PASS getComputedStyle(test, '').backgroundColor is transparent
PASS getComputedStyle(test, '').backgroundColor is black
PASS successfullyParsed is true

TEST COMPLETE
